kmeans-sklearn/requirements.txt
numpy==1.16.2
pillow==5.4.1
scikit-learn==0.20.3
scipy==1.2.1
sklearn==0.0
